Serial improvements and added a license

+++ drivers/attiny817_misc.h
@@ -9,7 +9,7 @@
 #define ATTINY817_MISC_H_
 
 void delay(uint32_t time_ms);
-long map(long x, long in_min, long in_max, long out_min, long out_max);
-unsigned long millis();
+long map(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max);
+uint32_t millis();
 
 #endif /* ATTINY817_MISC_H_ */
